@@ -18,9 +18,9 @@ package com.justwayward.reader.base;
 import android.app.Activity;
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.LayoutRes;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
+import androidx.annotation.LayoutRes;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentActivity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -101,10 +101,6 @@ public abstract class BaseFragment extends Fragment {
     public Context getApplicationContext() {
         return this.activity == null ? (getActivity() == null ? null : getActivity()
                 .getApplicationContext()) : this.activity.getApplicationContext();
-    }
-
-    protected LayoutInflater getLayoutInflater() {
-        return inflater;
     }
 
     protected View getParentView() {
